Function that adds a new element at the beginning of a list_t list

+++ 0x12-singly_linked_lists/2-add_node.c
@@ -20,7 +20,7 @@ list_t *add_node(list_t **head, const char *str)
 		return (free(new), NULL);
 
 	new->str = strdup(str);
-	new->len = strlen(str);
+	new->len = (int)strlen(str);
 	new->next = *head;
 	*head = new;
 	return (new);
